Publicador/requirements.txt
paho_mqtt==2.1.0
psutil==7.1.3
